+++ services/src/main/resources/META-INF/services/org.keycloak.wellknown.WellKnownProviderFactory
@@ -18,3 +18,4 @@
 org.keycloak.protocol.oidc.OIDCWellKnownProviderFactory
 org.keycloak.authorization.config.UmaWellKnownProviderFactory
 org.keycloak.protocol.oid4vc.issuance.OID4VCIssuerWellKnownProviderFactory
+org.keycloak.protocol.oauth2.OAuth2WellKnownProviderFactory
